requirements.txt
python3.5
pytorch0.4.1
cuda9.0
cudnn7.1.3
cython
cffi
opencv-python
scipy
msgpack
easydict
matplotlib
pyyaml
tensorboardX
